[add update]AuthContextを別ファイルに記述し、可読性向上
AuthContextの処理を一つのファイルに書くと冗長なので別ファイルに移しました。
